+++ lib/data/data.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class Data {
+   Data._();
+
+   static List<String> periodOfTime = ['Year', 'Month', 'Days'];
+   static const String image = "images/calculator.png";
+}
